+++ src/Eawv.Service/Models/UserModel.cs
@@ -17,5 +17,7 @@ public class UserModel
 
     public string Username { get; set; }
 
+    public string Email { get; set; }
+
     public List<PartyModel> Tenants { get; set; }
 }