+++ test/Eawv.Service.Integration.Tests/UserTests/DeleteUserAuthTest.cs
@@ -1,0 +1,46 @@
+// (c) Copyright by Abraxas Informatik AG
+// For license information see LICENSE file
+
+using System.Collections.Generic;
+using System.Net;
+using System.Net.Http;
+using System.Threading.Tasks;
+using Eawv.Service.Authentication;
+using Eawv.Service.Integration.Tests.MockedData;
+using Xunit;
+
+namespace Eawv.Service.Integration.Tests.UserTests;
+
+public class DeleteUserAuthTest : BaseRestTest
+{
+    private const string Url = "api/users/";
+
+    public DeleteUserAuthTest(TestApplicationFactory factory)
+        : base(factory)
+    {
+    }
+
+    public override async Task InitializeAsync()
+    {
+        await base.InitializeAsync();
+        await ElectionMockData.Seed(RunScoped);
+    }
+
+    [Fact]
+    public async Task TestAsElectionAdmin()
+    {
+        await AssertStatus(
+            () => ElectionAdminClient.DeleteAsync(Url + UserMockData.TestUser.Id),
+            HttpStatusCode.OK);
+    }
+
+    protected override IEnumerable<string> AuthorizedRoles()
+    {
+        yield return Role.Wahlverwalter;
+    }
+
+    protected override Task<HttpResponseMessage> AuthorizationTestCall(HttpClient httpClient)
+    {
+        return httpClient.DeleteAsync(Url + UserMockData.TestUser.Id);
+    }
+}